UPDATE: post users

+++ src/services/activity.ts
@@ -3,8 +3,8 @@ import {
 } from "@/lib/supabase/main";
 import type { Todo } from "@/types/todos";
 
-const getTodos = async () => {
-    const { data, error } = await supabase.from("activity").select("*");
+const getTodos = async ({ user_id }: { user_id: number }) => {
+    const { data, error } = await supabase.from("activity").select("*").eq("user_id", user_id);
 
     if (error) {
         throw new Error(error.message);
@@ -13,14 +13,15 @@ const getTodos = async () => {
     return data as Todo[];
 }
 
-const addTodo = async ({title, description, date} : {title: string, description?: string, date?: string}) => {
+const addTodo = async ({ user_id, title, description, date }: { user_id: number, title: string, description?: string, date?: string }) => {
     const { error } = await supabase
         .from('activity')
-        .insert({ 
+        .insert({
             title: title,
             description: description,
-            schedule: date
-         });
+            schedule: date,
+            user_id: user_id
+        });
 
     if (error) {
         throw new Error(error.message);
@@ -31,23 +32,25 @@ const addTodo = async ({title, description, date} : {title: string, description?
 
 const updateTodo = async (
     id: number,
+    user_id: number,
     updates: { title?: string; description?: string; schedule?: string; is_completed?: boolean }
 ) => {
     const { error } = await supabase
         .from("activity")
         .update(updates)
-        .eq("id", id);
+        .eq("id", id)
+        .eq("user_id", user_id);
 
     if (error) throw new Error(error.message);
     return true;
 };
 
-
-const deleteTodos = async (selectedIds: number) => {
+const deleteTodos = async (selectedIds: number, user_id: number) => {
     const { error } = await supabase
         .from("activity")
         .delete()
-        .eq("id", selectedIds);
+        .eq("id", selectedIds)
+        .eq("user_id", user_id);
 
     if (error) {
         throw new Error(error.message);
